Move Accuracy method
Move from cnn.py -> model_helper as requested by @akyrola
